+++ GenerateEverything.hs
@@ -59,7 +59,7 @@ unsafeModules = map modToFile
   , "System.Exit.Primitive"
   , "Text.Pretty.Core"
   , "Text.Pretty"
-  ]
+  ] ++ sizedTypesModules
 
 isUnsafeModule :: FilePath -> Bool
 isUnsafeModule fp =
@@ -97,6 +97,55 @@ isWithKModule =
   -- to `isWithKModule`.
   \ fp -> unqualifiedModuleName fp == "WithK"
        || fp `elem` withKModules
+
+sizedTypesModules :: [FilePath]
+sizedTypesModules = map modToFile
+  [ "Codata.Cofin"
+  , "Codata.Cofin.Literals"
+  , "Codata.Colist"
+  , "Codata.Colist.Bisimilarity"
+  , "Codata.Colist.Categorical"
+  , "Codata.Colist.Properties"
+  , "Codata.Conat"
+  , "Codata.Conat.Bisimilarity"
+  , "Codata.Conat.Literals"
+  , "Codata.Conat.Properties"
+  , "Codata.Covec"
+  , "Codata.Covec.Bisimilarity"
+  , "Codata.Covec.Categorical"
+  , "Codata.Covec.Instances"
+  , "Codata.Covec.Properties"
+  , "Codata.Cowriter"
+  , "Codata.Cowriter.Bisimilarity"
+  , "Codata.Delay"
+  , "Codata.Delay.Bisimilarity"
+  , "Codata.Delay.Categorical"
+  , "Codata.Delay.Properties"
+  , "Codata.M"
+  , "Codata.M.Bisimilarity"
+  , "Codata.M.Properties"
+  , "Codata.Stream"
+  , "Codata.Stream.Bisimilarity"
+  , "Codata.Stream.Categorical"
+  , "Codata.Stream.Instances"
+  , "Codata.Stream.Properties"
+  , "Codata.Thunk"
+  , "Data.Container"
+  , "Data.Container.Any"
+  , "Data.Container.FreeMonad"
+  , "Data.Nat.PseudoRandom.LCG.Unsafe"
+  , "Data.Tree.Rose"
+  , "Data.Tree.Rose.Properties"
+  , "Data.Trie"
+  , "Data.Trie.NonEmpty"
+  , "Relation.Unary.Sized"
+  , "Size"
+  , "Text.Tree.Linear"
+  ]
+
+isSizedTypesModule :: FilePath -> Bool
+isSizedTypesModule =
+  \ fp -> fp `elem` sizedTypesModules
 
 unqualifiedModuleName :: FilePath -> String
 unqualifiedModuleName = dropExtension . takeFileName
@@ -139,10 +188,9 @@ extractHeader mod = extract
                               , "Please see other existing files or consult HACKING.md."
                               ]
 
--- | A crude classifier looking for lines containing options & trying to guess
---   whether the safe file is using either @--guardedness@ or @--sized-types@
+-- | A crude classifier looking for lines containing options
 
-data Status = Deprecated | Unsafe | Safe | SafeGuardedness | SafeSizedTypes
+data Status = Deprecated | Unsafe | Safe
   deriving (Eq)
 
 classify :: FilePath -> [String] -> [String] -> Status
@@ -156,8 +204,6 @@ classify fp hd ls
   -- And then perform the actual classification
   | deprecated                = Deprecated
   | isUnsafe                  = Unsafe
-  | guardedness               = SafeGuardedness
-  | sizedtypes                = SafeSizedTypes
   | safe                      = Safe
   -- We know that @not (isUnsafe || safe)@, all cases are covered
   | otherwise                 = error "IMPOSSIBLE"
@@ -169,8 +215,6 @@ classify fp hd ls
     isUnsafe = isUnsafeModule fp
 
     -- based on detected OPTIONS
-    guardedness = option "--guardedness"
-    sizedtypes  = option "--sized-types"
     safe        = option "--safe"
     withK       = option "--with-K"
     withoutK    = option "--without-K"
@@ -223,9 +267,7 @@ checkFilePaths cat fps = forM_ fps $ \ fp -> do
 -- Collecting all non-Core library files, analysing them and generating
 -- 4 files:
 -- Everything.agda                 all the modules
--- EverythingSafe.agda             all the safe modules (may be incompatible)
--- EverythingSafeGuardedness.agda  all the safe modules using --guardedness
--- EverythingSafeSizedTypes.agda   all the safe modules using --sized-types
+-- EverythingSafe.agda             all the safe modules
 
 main = do
   args <- getArgs
@@ -254,25 +296,9 @@ main = do
 
   writeFileUTF8 (safeOutputFile ++ ".agda") $
     unlines [ header
-            , "{-# OPTIONS --guardedness --sized-types #-}\n"
+            , "{-# OPTIONS --safe --guardedness #-}\n"
             , mkModule safeOutputFile
             , format $ filter ((Unsafe /=) . status) libraryfiles
-            ]
-
-  let safeGuardednessOutputFile = safeOutputFile ++ "Guardedness"
-  writeFileUTF8 (safeGuardednessOutputFile ++ ".agda") $
-    unlines [ header
-            , "{-# OPTIONS --safe --guardedness #-}\n"
-            , mkModule safeGuardednessOutputFile
-            , format $ filter ((SafeGuardedness ==) . status) libraryfiles
-            ]
-
-  let safeSizedTypesOutputFile = safeOutputFile ++ "SizedTypes"
-  writeFileUTF8 (safeSizedTypesOutputFile ++ ".agda") $
-    unlines [ header
-            , "{-# OPTIONS --safe --sized-types #-}\n"
-            , mkModule safeSizedTypesOutputFile
-            , format $ filter ((SafeSizedTypes ==) . status) libraryfiles
             ]
 
 -- | Usage info.